+++ src/test/java/org/folio/sidecar/it/ForwardUnknownEgressIT.java
@@ -25,7 +25,6 @@ import org.junit.jupiter.api.Test;
 
 @IntegrationTest
 @TestProfile(ForwardUnknownTestProfile.class)
-//@QuarkusTestResource(WireMockExtension.class)
 @EnableWireMock(verbose = true)
 class ForwardUnknownEgressIT {
